[CODE] Make buildable pt 2
