+++ src/react-app/SimulationDrawer.tsx
@@ -1,8 +1,6 @@
 import { 
     animated, 
     useSpring, 
-    useTransition,
-    config,
 } from '@react-spring/web';
 
 // import {type AppNode } from './ProteinNode';
